Move excess logic from dbHelper.
